MR Changes
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@5668 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/hydrolibs/hydroqguielementutil/paintutils.cpp
@@ -197,7 +197,7 @@ paintCovarianceEllipse( QPainter *p, float pxx, float pxy, float pyy )
     cov.ellipse( a, b, psi );
 
     QPointF covPoint(a,b);
-    float psiDeg = (int)ceil( psi *180/M_PI );            // rot. angle into princ. axes
+    float psiDeg = ceil( psi *180/M_PI );            // rot. angle into princ. axes
 
     // Paint uncertainty ellipse
     p->save();
